% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockfile_mod.R
\name{lockfileUI}
\alias{lockfileUI}
\alias{lockfileServer}
\title{Lockfile Download Module}
\usage{
lockfileUI(id)

lockfileServer(id)
}
\arguments{
\item{id}{An ID string that identifies the namespace of the UI and server components.}
}
\description{
A module that is included in the navbar dropdown that allows the application user
to download the lockfile of the relevant packages of the reproducible reactives.
}
