% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro.R
\name{repro}
\alias{repro}
\title{Reproduce Code}
\usage{
repro(x, ..., env = rlang::caller_env())
}
\arguments{
\item{x}{\code{\link[shiny]{reactive}} object to make reproducible}

\item{...}{Additional arguments to pass to other methods}

\item{env}{The environment `x` is defined in. By default it is the environment of where `repro` is called}
}
\description{
Reproduce Code
}
