% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_mod.R
\name{reactiveTabUI}
\alias{reactiveTabUI}
\title{Reproducible Code Modules}
\usage{
reactiveTabUI(id)
}
\arguments{
\item{id}{The ID string to be namespaced}
}
\description{
A series of modules that are used
}
\details{
\code{reactiveTabUI} -
}
