% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro_S7.R
\name{Repro}
\alias{Repro}
\title{Reproduce Code}
\usage{
Repro(code = list(), packages = character(0), prerequesites = list())
}
\arguments{
\item{code}{Code chunks found in a given expression}

\item{packages}{Packages found in the function calls in the code and/or pre-requisites}

\item{prerequesites}{Code chunks used to generate reactive objects found in the code}
}
\description{
Reproduce Code
}
