% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro_tab_mod.R
\name{repro_tab_ui}
\alias{repro_tab_ui}
\title{Reproducible Tab Module}
\usage{
repro_tab_ui(id, title, description, filters = NULL)
}
\arguments{
\item{id}{}
}
\description{
A generic module that can show reproducible code in the left hand side with
the module code in the right hand side.
}
