% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro.R
\name{repro}
\alias{repro}
\title{Reproduce Code}
\usage{
repro(x, ..., env = rlang::caller_env())
}
\arguments{
\item{x}{Reactive object}

\item{...}{Additional arguments passed to methods}

\item{env}{The environment the}
}
\description{
Reproduce Code
}
