% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_mod.R, R/ifelse_mod.R, R/input_mod.R,
%   R/multi_level_mod.R, R/passed_reactive_mod.R, R/reactive_mod.R,
%   R/reactive_val_mod.R, R/reactive_values_mod.R
\name{bindTabUI}
\alias{bindTabUI}
\alias{ifelseTabUI}
\alias{inputTabUI}
\alias{multiLevelModuleUI}
\alias{passedReactiveTabUI}
\alias{reactiveTabUI}
\alias{reactiveValTabUI}
\alias{reactiveValuesTabUI}
\title{Reproducible Code Modules}
\usage{
bindTabUI(id)

ifelseTabUI(id)

inputTabUI(id)

multiLevelModuleUI(id)

passedReactiveTabUI(id)

reactiveTabUI(id)

reactiveValTabUI(id)

reactiveValuesTabUI(id)
}
\arguments{
\item{id}{An ID string that identifies the namespace of the UI and server components.}
}
\description{
A series of modules that are used to show how the reproducing of Shiny code works.
}
\details{
\code{bindTabUI} - \code{\link[shiny]{bindEvent}} and \code{\link[shiny]{bindCache}} being used within a reactive

\code{ifelseTabUI} - An if/else statement included in a reactive, showing it only returns relevant code
in the reproducing section

\code{inputTabUI} - A Shiny input being used within the reactive

\code{multiLevelModuleUI} - Reactives being passed to a sub-module, only to be rendered in the parent module

\code{passedReactiveTabUI} - A reactive from another module being used in the selected module

\code{reactiveTabUI} - A reactive object being passed into another reactive object

\code{reactiveValuesTabUI} - A \code{\link[shiny]{reactiveVal}} used within a reactive

\code{reactiveValuesTabUI} - An element in \code{\link[shiny]{reactiveValues}} used within a reactive
}
